added txid

+++ scripts/init_libeufin_sandbox.sh
@@ -213,9 +213,6 @@ if [ ! -f "/app/initdone" ]; then
 
 fi
 
-touch /app/initdone
-ls -la /app/initdone
-
 #  install versions according to LibFinEu/frontend/README.md
 
 echo list ebicssubscriber
@@ -228,42 +225,35 @@ libeufin-cli connections list-connections
 echo list show-connection
 libeufin-cli connections show-connection $CONNECTION_NAME  
 
+if [ ! -f "/app/initdone" ]; then
+    apt update --allow-releaseinfo-change
+    apt-get install -y jq qpdf xxd libxml2-utils openssl
 
-apt update --allow-releaseinfo-change
-apt-get install -y jq qpdf xxd libxml2-utils openssl
+    # client_pr_key="${CLIENT_PR_KEY:-/app/scripts/client_private_key.pem}"
+    # client_pub_key="${CLIENT_PUB_KEY:-/app/scripts/client_public_key.pem}"
+    # cat /app/scripts/backupfile | jq -r '.sigBlob' | openssl enc -d -base64 -A | openssl pkcs8 -inform DER -outform PEM -out $client_pr_key  -passin pass:$SECRET
+    # openssl rsa -pubout -in $client_pr_key -out $client_pub_key
+    # echo "client pk exported to $client_pr_key public key to $client_pub_key "
+    client_pr_key=${CLIENT_PR_KEY_OUT:-/app/keys/client_private_key.pem}
+    client_pub_key=${CLIENT_PUB_KEY_OUT:-/app/keys/client_public_key.pem}
+    bank_pub_key=${BANK_PUB_KEY_OUT:-/app/keys/bank_public_key.pem}
 
-# client_pr_key="${CLIENT_PR_KEY:-/app/scripts/client_private_key.pem}"
-# client_pub_key="${CLIENT_PUB_KEY:-/app/scripts/client_public_key.pem}"
-# cat /app/scripts/backupfile | jq -r '.sigBlob' | openssl enc -d -base64 -A | openssl pkcs8 -inform DER -outform PEM -out $client_pr_key  -passin pass:$SECRET
-# openssl rsa -pubout -in $client_pr_key -out $client_pub_key
-# echo "client pk exported to $client_pr_key public key to $client_pub_key "
-client_pr_key=${CLIENT_PR_KEY_OUT:-/app/keys/client_private_key.pem}
-client_pub_key=${CLIENT_PUB_KEY_OUT:-/app/keys/client_public_key.pem}
-bank_pub_key=${BANK_PUB_KEY_OUT:-/app/keys/bank_public_key.pem}
+    echo "save keys to $client_pr_key $client_pub_key $bank_pub_key "
+    sql="SELECT \"encryptionPrivateKey\" from nexusebicssubscribers s where \"hostID\"='testhost'"
+    # read hex key  | convert to binary | create pem
+    PGPASSWORD=$POSTGRES_PASSWORD psql -d "$POSTGRES_DB" -h "$POSTGRES_HOST" -U "$POSTGRES_USER" -t -c "$sql" | xxd -r -p | openssl pkcs8 -topk8 -nocrypt -inform DER -out $client_pr_key
+    openssl rsa -in $client_pr_key -pubout -out $client_pub_key
+    echo "get bank pub auth key"
+    sql="SELECT \"bankAuthenticationPublicKey\" from nexusebicssubscribers s where \"hostID\"='testhost'"
+    PGPASSWORD=$POSTGRES_PASSWORD psql -d "$POSTGRES_DB" -h "$POSTGRES_HOST" -U "$POSTGRES_USER" -t -c "$sql" | xxd -r -p | openssl rsa -pubin -inform DER -outform PEM -out $bank_pub_key
+    sql="SELECT \"bankEncryptionPublicKey\" from nexusebicssubscribers s where \"hostID\"='testhost'"
+    PGPASSWORD=$POSTGRES_PASSWORD psql -d "$POSTGRES_DB" -h "$POSTGRES_HOST" -U "$POSTGRES_USER" -t -c "$sql" | xxd -r -p | openssl rsa -pubin -inform DER -outform PEM -out $bank_pub_key-enc
 
-echo "save keys to $client_pr_key $client_pub_key $bank_pub_key "
-sql="SELECT \"encryptionPrivateKey\" from nexusebicssubscribers s where \"hostID\"='testhost'"
-# read hex key  | convert to binary | create pem
-PGPASSWORD=$POSTGRES_PASSWORD psql -d "$POSTGRES_DB" -h "$POSTGRES_HOST" -U "$POSTGRES_USER" -t -c "$sql" | xxd -r -p | openssl pkcs8 -topk8 -nocrypt -inform DER -out $client_pr_key
-openssl rsa -in $client_pr_key -pubout -out $client_pub_key
-echo "get bank pub auth key"
-sql="SELECT \"bankAuthenticationPublicKey\" from nexusebicssubscribers s where \"hostID\"='testhost'"
-PGPASSWORD=$POSTGRES_PASSWORD psql -d "$POSTGRES_DB" -h "$POSTGRES_HOST" -U "$POSTGRES_USER" -t -c "$sql" | xxd -r -p | openssl rsa -pubin -inform DER -outform PEM -out $bank_pub_key
-sql="SELECT \"bankEncryptionPublicKey\" from nexusebicssubscribers s where \"hostID\"='testhost'"
-PGPASSWORD=$POSTGRES_PASSWORD psql -d "$POSTGRES_DB" -h "$POSTGRES_HOST" -U "$POSTGRES_USER" -t -c "$sql" | xxd -r -p | openssl rsa -pubin -inform DER -outform PEM -out $bank_pub_key-enc
+    echo "key generation done"
+fi
 
-echo "key generation done"
-# alternatively read keys from backup file
-# apt-get install jq -y # qpdf xxd libxml2-utils openssl -y
-# cat /app/scripts/backupfile | jq -r '.sigBlob' | openssl enc -d -base64 -A | openssl pkcs8 -inform DER -outform PEM -out $client_pr_key  -passin pass:$SECRET
-# openssl rsa -pubout -in $client_pr_key -out $client_pub_key
-# echo "client pk exported to $client_pr_key public key to $client_pub_key "
-
-
-# libeufin-cli accounts task-schedule --task-name=fetch-statement --task-type=fetch --task-cronspec="30 * * * *" --task-param-range-type=all --task-param-level=statement  $IBAN
-# libeufin-cli accounts task-schedule --task-name=fetch-statement --task-type=fetch --task-cronspec="30 * * * *" --task-param-range-type=all --task-param-level=statement  $EXTERNAL_IBAN
-# libeufin-cli accounts task-schedule --task-name=fetch-statement --task-type=fetch --task-cronspec="30 * * * *" --task-param-range-type=all --task-param-level=statement  $REGISTERED_IBAN
-# echo " auto fetch registered"
+touch /app/initdone
+ls -la /app/initdone
 
 read -t 10 -p "Setup & startup of nexus and sandbox complete, starting Libeufin react-ui UI on localhost:3000, login with:  $LIBEUFIN_NEXUS_USERNAME $LIBEUFIN_NEXUS_PASSWORD " || true
 #serve -s build
